Подготовил тестовый проект в пакете example2, для дальнейшего изучения инъекции зависимостей
