--- app/src/main/java/ru/paramonov/startinjectdepenency/example2/data/ExampleDatabase.kt
@@ -1,8 +1,0 @@
-package ru.paramonov.startinjectdepenency.example2.data
-
-class ExampleDatabase {
-
-    fun method() {
-
-    }
-}
