--- app/src/main/java/ru/paramonov/startinjectdepenency/example2/data/ExampleLocalDataSource.kt
@@ -1,6 +1,0 @@
-package ru.paramonov.startinjectdepenency.example2.data
-
-interface ExampleLocalDataSource {
-
-    fun method()
-}
