--- app/src/main/java/ru/paramonov/startinjectdepenency/example2/data/ExampleMapper.kt
@@ -1,8 +1,0 @@
-package ru.paramonov.startinjectdepenency.example2.data
-
-class ExampleMapper {
-
-    fun map() {
-
-    }
-}
