--- app/src/main/java/ru/paramonov/startinjectdepenency/example2/data/ExampleRemoteDataSource.kt
@@ -1,6 +1,0 @@
-package ru.paramonov.startinjectdepenency.example2.data
-
-interface ExampleRemoteDataSource {
-
-    fun method()
-}
