--- app/src/main/java/ru/paramonov/startinjectdepenency/example2/data/ExampleRemoteDataSourceImpl.kt
@@ -1,10 +1,0 @@
-package ru.paramonov.startinjectdepenency.example2.data
-
-class ExampleRemoteDataSourceImpl(
-    private val apiService: ExampleApiService
-) : ExampleRemoteDataSource {
-
-    override fun method() {
-
-    }
-}
