--- app/src/main/java/ru/paramonov/startinjectdepenency/example2/data/ExampleRepositoryImpl.kt
@@ -1,13 +1,0 @@
-package ru.paramonov.startinjectdepenency.example2.data
-
-import ru.paramonov.startinjectdepenency.example2.domain.ExampleRepository
-
-class ExampleRepositoryImpl(
-    private val localDataSource: ExampleLocalDataSource,
-    private val remoteDataSource: ExampleRemoteDataSource
-) : ExampleRepository {
-
-    override fun method() {
-
-    }
-}
