+++ app/src/main/java/ru/paramonov/startinjectdepenency/example2/data/database/ExampleDatabase.kt
@@ -1,0 +1,15 @@
+package ru.paramonov.startinjectdepenency.example2.data.database
+
+import android.util.Log
+
+class ExampleDatabase {
+
+    fun method() {
+        Log.d(LOG_TAG, "ExampleDatabase")
+    }
+
+    companion object {
+
+        private const val LOG_TAG = "EXAMPLE_TEST"
+    }
+}
