+++ app/src/main/java/ru/paramonov/startinjectdepenency/example2/data/datasource/ExampleRemoteDataSource.kt
@@ -1,0 +1,6 @@
+package ru.paramonov.startinjectdepenency.example2.data.datasource
+
+interface ExampleRemoteDataSource {
+
+    fun method()
+}
