+++ app/src/main/java/ru/paramonov/startinjectdepenency/example2/data/datasource/ExampleRemoteDataSourceImpl.kt
@@ -1,0 +1,12 @@
+package ru.paramonov.startinjectdepenency.example2.data.datasource
+
+import ru.paramonov.startinjectdepenency.example2.data.network.ExampleApiService
+
+class ExampleRemoteDataSourceImpl(
+    private val apiService: ExampleApiService
+) : ExampleRemoteDataSource {
+
+    override fun method() {
+        apiService.method()
+    }
+}
