+++ app/src/main/java/ru/paramonov/startinjectdepenency/example2/data/mapper/ExampleMapper.kt
@@ -1,0 +1,8 @@
+package ru.paramonov.startinjectdepenency.example2.data.mapper
+
+class ExampleMapper {
+
+    fun map() {
+
+    }
+}
