+++ app/src/main/java/ru/paramonov/startinjectdepenency/example2/data/network/ExampleApiService.kt
@@ -1,0 +1,15 @@
+package ru.paramonov.startinjectdepenency.example2.data.network
+
+import android.util.Log
+
+class ExampleApiService {
+
+    fun method() {
+        Log.d(LOG_TAG, "ExampleApiService")
+    }
+
+    companion object {
+
+        private const val LOG_TAG = "EXAMPLE_TEST"
+    }
+}
