+++ app/src/main/java/ru/paramonov/startinjectdepenency/example2/domain/ExampleUseCase.kt
@@ -7,6 +7,6 @@ class ExampleUseCase(
 ) {
 
     operator fun invoke() {
-
+        repository.method()
     }
 }
