+++ app/src/main/java/ru/paramonov/startinjectdepenency/example2/presentation/ExampleViewModel.kt
@@ -7,6 +7,6 @@ class ExampleViewModel(
 ) {
 
     fun method() {
-
+        useCase
     }
 }
